requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
python-dotenv
requests
authlib>=0.14.1
six